[BE][Feat] elasticsearch를 이용한 단건 update 구현, 소비자 가게 목록 조회 elasticsearch 적용

See merge request s10-final/S10P31D104!171

+++ backend/src/main/java/com/todayeat/backend/store/dto/response/GetConsumerListStoreResponse.java
@@ -48,13 +48,13 @@ public class GetConsumerListStoreResponse {
         private String operatingTime;
 
         @Schema(description = "리뷰 수", example = "0")
-        private int reviewCnt;
+        private Integer reviewCnt;
 
         @Schema(description = "찜 수", example = "0")
-        private int favoriteCnt;
+        private Integer favoriteCnt;
 
         @Schema(description = "해당 위치와 거리", example = "1500")
-        private Double distance;
+        private Integer distance;
 
         @Schema(description = "카테고리 목록", example = "{name: 카테고리1, image: https://todayeat-bucket.s3.ap-northeast-2.amazonaws.com/seller/1/store-image/img1.png, name: 카테고리2, image: https://todayeat-bucket.s3.ap-northeast-2.amazonaws.com/seller/1/store-image/img2.png}")
         private List<CategoryInfo> categoryList;
